backend/requirements.txt
fastapi
uvicorn
pdfplumber
scikit-learn
sentence-transformers
slowapi
python-multipart
psycopg2-binary
python-jose[cryptography]
pydantic[email]
python-dotenv
prisma
supabase
pydantic-settings

